Remove static methods and wrap in SceneGraph class
