+++ tesseract_scene_graph/include/tesseract_scene_graph/graph.h
@@ -29,6 +29,10 @@
 #include <boost/graph/adjacency_list.hpp> // for customizable graphs
 #include <boost/graph/directed_graph.hpp> // A subclass to provide reasonable arguments to adjacency_list for a typical directed graph
 #include <boost/graph/properties.hpp>
+#include <boost/graph/dijkstra_shortest_paths.hpp>
+#include <boost/graph/depth_first_search.hpp>
+#include <fstream>
+//#include <boost/graph/copy.hpp>
 #include <string>
 #include <unordered_map>
 #include <tesseract_scene_graph/link.h>
@@ -73,199 +77,360 @@ typedef boost::property<boost::edge_joint_t, JointConstPtr,
 
 
 typedef boost::adjacency_list<boost::vecS, boost::vecS, boost::bidirectionalS, VertexProperty, EdgeProperty, GraphProperty> Graph;
-typedef std::shared_ptr<Graph> GraphPtr;
-typedef std::shared_ptr<const Graph> GraphConstPtr;
-
-typedef Graph::vertex_descriptor Vertex;
-typedef Graph::edge_descriptor Edge;
-
-/**
- * @brief Sets the graph name
- * @param name The name of the graph
- * @param graph The graph
- */
-static inline void setName(const std::string& name, Graph& graph)
+class SceneGraph : public Graph
 {
-  boost::get_property(graph, boost::graph_name) = name;
-}
+public:
 
-/**
- * @brief Sets the graph name
- * @param name The name of the graph
- * @param graph The graph
- */
-static inline const std::string& getName(const Graph& graph)
+  /**
+   * @brief Holds the shortest path information.
+   *
+   * The first vector is a list of links along the shortest path
+   * The second vector is a list of joints along the shortest path
+   */
+  typedef std::pair<std::vector<std::string>, std::vector<std::string>> Path;
+  typedef SceneGraph::vertex_descriptor Vertex;
+  typedef SceneGraph::edge_descriptor Edge;
+
+  /**
+   * @brief Sets the graph name
+   * @param name The name of the graph
+   */
+  void setName(const std::string& name)
+  {
+    boost::set_property(static_cast<Graph&>(*this), boost::graph_name, name);
+  }
+
+  /**
+   * @brief Sets the graph name
+   * @param name The name of the graph
+   */
+  const std::string& getName() const
+  {
+    return boost::get_property(static_cast<const Graph&>(*this), boost::graph_name);
+  }
+
+  /**
+   * @brief Sets the root link name (aka. World Coordinate Frame)
+   * @param name The name of the link
+   * @return Return False if a link does not exists, otherwise true
+   */
+  bool setRoot(const std::string& name)
+  {
+    auto& map = get_property(static_cast<const Graph&>(*this), boost::graph_link_map);
+    auto found = map.find(name);
+
+    if (found != map.end())
+      return false;
+
+    boost::set_property(static_cast<Graph&>(*this), boost::graph_root, name);
+
+    return true;
+  }
+
+  /**
+   * @brief Gets the root link name (aka. World Coordinate Frame)
+   * @return The root link name
+   */
+  const std::string& getRoot() const
+  {
+    return boost::get_property(static_cast<const Graph&>(*this), boost::graph_root);
+  }
+
+  /**
+   * @brief Adds a link to the graph
+   * @param link The link to be added to the graph
+   * @return Return False if a link with the same name allready exists, otherwise true
+   */
+  bool addLink(LinkPtr link)
+  {
+    auto& map = boost::get_property(static_cast<Graph&>(*this), boost::graph_link_map);
+    auto found = map.find(link->getName());
+
+    if (found != map.end())
+      return false;
+
+    VertexProperty info(link);
+    Vertex v = boost::add_vertex(info, static_cast<Graph&>(*this));
+    map[link->getName()] = std::make_pair(link, v);
+    return true;
+  }
+
+  /**
+   * @brief Get a link in the graph
+   * @param name The name of the link
+   * @return Return nullptr if link name does not exists, otherwise a pointer to the link
+   */
+  LinkConstPtr getLink(const std::string& name) const
+  {
+    auto& map = boost::get_property(static_cast<const Graph&>(*this), boost::graph_link_map);
+    auto found = map.find(name);
+
+    if (found == map.end())
+      return nullptr;
+
+    return found->second.first;
+  }
+
+  /**
+   * @brief Removes a link from the graph
+   * @param name Name of the link to be removed
+   * @return Return False if a link does not exists, otherwise true
+   */
+  bool removeLink(const std::string& name)
+  {
+    auto& map = boost::get_property(static_cast<Graph&>(*this), boost::graph_link_map);
+    auto found = map.find(name);
+
+    if (found == map.end())
+      return false;
+
+    boost::remove_vertex(found->second.second, static_cast<Graph&>(*this));
+    map.erase(name);
+
+    return true;
+  }
+
+  /**
+   * @brief Adds joint to the graph
+   * @param joint The joint to be added
+   * @return Return False if parent or child link does not exists and if joint name already exists in the graph, otherwise true
+   */
+  bool addJoint(JointPtr joint)
+  {
+    auto& link_map = boost::get_property(static_cast<Graph&>(*this), boost::graph_link_map);
+    auto& joint_map = boost::get_property(static_cast<Graph&>(*this), boost::graph_joint_map);
+    auto parent = link_map.find(joint->parent_link_name);
+    auto child = link_map.find(joint->child_link_name);
+    auto found = joint_map.find(joint->getName());
+
+    if ( (parent == link_map.end()) || (child == link_map.end()) || (found != joint_map.end()) )
+      return false;
+
+    double d = joint->parent_to_joint_origin_transform.translation().norm();
+    EdgeProperty info(joint, d);
+    std::pair<Edge, bool> e = boost::add_edge(parent->second.second, child->second.second, info, static_cast<Graph&>(*this));
+    assert(e.second == true);
+    joint_map[joint->getName()] = std::make_pair(joint, e.first);
+
+    return true;
+  }
+
+  /**
+   * @brief Get a joint in the graph
+   * @param name The name of the joint
+   * @return Return nullptr if joint name does not exists, otherwise a pointer to the joint
+   */
+  JointConstPtr getJoint(const std::string& name) const
+  {
+    auto& map = boost::get_property(static_cast<const Graph&>(*this), boost::graph_joint_map);
+    auto found = map.find(name);
+
+    if (found == map.end())
+      return nullptr;
+
+    return found->second.first;
+  }
+
+  /**
+   * @brief Removes a joint from the graph
+   * @param name Name of the joint to be removed
+   * @return Return False if a joint does not exists, otherwise true
+   */
+  bool removeJoint(const std::string& name)
+  {
+    auto& map = boost::get_property(static_cast<Graph&>(*this), boost::graph_joint_map);
+    auto found = map.find(name);
+
+    if (found == map.end())
+      return false;
+
+    boost::remove_edge(found->second.second, static_cast<Graph&>(*this));
+    map.erase(name);
+
+    return true;
+  }
+
+  /**
+   * @brief Determine if the graph contains cycles
+   * @return True if graph is acyclic otherwise false
+   */
+  bool isAcyclic() const
+  {
+    bool acyclic = true;
+    cycle_detector vis(acyclic);
+    boost::depth_first_search(static_cast<const Graph&>(*this), boost::visitor(vis));
+    return acyclic;
+  }
+
+  /**
+   * @brief Determine if the graph is a tree
+   * @return True if graph is tree otherwise false
+   */
+  bool isTree() const
+  {
+    bool tree = true;
+    tree_detector vis(tree);
+    boost::depth_first_search(static_cast<const Graph&>(*this), boost::visitor(vis));
+    return tree;
+  }
+
+  /**
+   * @brief Saves Graph as Graph Description Language (DOT)
+   * @param path The file path
+   */
+  void saveDOT(std::string path)
+  {
+    std::ofstream dot_file(path);
+
+    dot_file << "digraph D {\n"
+      << "  rankdir=LR\n"
+      << "  size=\"4,3\"\n"
+      << "  ratio=\"fill\"\n"
+      << "  edge[style=\"bold\"]\n" << "  node[shape=\"circle\"]\n";
+
+    const Graph& graph = static_cast<const Graph&>(*this);
+    Graph::edge_iterator ei, ei_end;
+    for (boost::tie(ei, ei_end) = boost::edges(graph); ei != ei_end; ++ei)
+    {
+      Edge e = *ei;
+      Vertex u = boost::source(e, graph);
+      Vertex v = boost::target(e, graph);
+      JointConstPtr joint = boost::get(boost::edge_joint, graph)[e];
+
+      dot_file << boost::get(boost::vertex_link, graph)[u]->getName()
+               << " -> " << boost::get(boost::vertex_link, graph)[v]->getName()
+               << "[label=\"" << joint->getName() << "\n(" << joint->type << ")\", color=\"black\"]";
+
+    }
+    dot_file << "}";
+  }
+
+  Path getShortestPath(const std::string& root, const std::string& tip)
+  {
+    const Graph& graph = static_cast<const Graph&>(*this);
+    std::vector<Vertex> p(boost::num_vertices(graph));
+    std::vector<double> d(boost::num_vertices(graph));
+    Vertex s = getVertex(root);
+    dijkstra_shortest_paths(graph, s, boost::predecessor_map(&p[0]).distance_map(&d[0]));
+
+    std::vector<std::string> links;
+    std::vector<std::string> joints;
+    Vertex v = getVertex(tip); // We want to start at the destination and work our way back to the source
+    for(Vertex u = p[v]; // Start by setting 'u' to the destintaion node's predecessor
+        u != v; // Keep tracking the path until we get to the source
+        v = u, u = p[v]) // Set the current vertex to the current predecessor, and the predecessor to one level up
+    {
+      links.push_back(boost::get(boost::vertex_link, graph)[v]->getName());
+      joints.push_back(boost::get(boost::edge_joint, graph)[boost::edge(u, v, graph).first]->getName());
+    }
+    links.push_back(root);
+    std::reverse(links.begin(), links.end());
+    std::reverse(joints.begin(), joints.end());
+
+  #ifdef NDEBUG
+    std::cout << "distances and parents:" << std::endl;
+    Graph::vertex_iterator vi, vend;
+    for (boost::tie(vi, vend) = boost::vertices(g); vi != vend; ++vi)
+    {
+      std::cout << "distance(" << boost::get(boost::vertex_link, g)[*vi]->getName() << ") = " << d[*vi] << ", ";
+      std::cout << "parent(" << boost::get(boost::vertex_link, g)[*vi]->getName() << ") = " << boost::get(boost::vertex_link, g)[p[*vi]]->getName() << std::endl;
+    }
+    std::cout << std::endl;
+  #endif
+    return Path(links, joints);
+  }
+
+  //static inline Graph copyGraph(const Graph& graph)
+  //{
+  //  Graph new_graph;
+  //  boost::copy_graph(graph, new_graph);
+  //  return new_graph;
+  //}
+
+private:
+  Vertex getVertex(const std::string& name) const
+  {
+    auto& map = boost::get_property(static_cast<const Graph&>(*this), boost::graph_link_map);
+    auto found = map.find(name);
+
+    if (found == map.end())
+      return Vertex();
+
+    return found->second.second;
+  }
+
+  Edge getEdge(const std::string& name) const
+  {
+    auto& map = boost::get_property(static_cast<const Graph&>(*this), boost::graph_joint_map);
+    auto found = map.find(name);
+
+    if (found == map.end())
+      return Edge();
+
+    return found->second.second;
+  }
+
+  struct cycle_detector : public boost::dfs_visitor<>
+  {
+    cycle_detector( bool& ascyclic)
+      : ascyclic_(ascyclic) { }
+
+    template <class e, class g>
+    void back_edge(e, g&) {
+      ascyclic_ = false;
+    }
+  protected:
+    bool& ascyclic_;
+  };
+
+  struct tree_detector : public boost::dfs_visitor<>
+  {
+    tree_detector( bool& tree)
+      : tree_(tree) { }
+
+    template <class u, class g>
+    void discover_vertex(u vertex, g graph)
+    {
+      int num_in_edges = boost::in_degree(vertex, graph);
+
+      if (num_in_edges > 1)
+      {
+        tree_ = false;
+        return;
+      }
+
+      // Check if not vertex is unused.
+      if (num_in_edges == 0 && boost::out_degree(vertex, graph) == 0)
+      {
+        tree_ = false;
+        return;
+      }
+    }
+
+    template <class e, class g>
+    void back_edge(e, g&)
+    {
+      tree_ = false;
+    }
+  protected:
+    bool& tree_;
+  };
+
+};
+typedef std::shared_ptr<SceneGraph> SceneGraphPtr;
+typedef std::shared_ptr<const SceneGraph> SceneGraphConstPtr;
+
+inline std::ostream& operator<<(std::ostream& os, const SceneGraph::Path& path)
 {
-  return boost::get_property(graph, boost::graph_name);
-}
+  os << "Links:" << std::endl;
+  for (auto l : path.first)
+    os << "  " << l << std::endl;
 
-/**
- * @brief Sets the root link name (aka. World Coordinate Frame)
- * @param name The name of the link
- * @param graph The graph
- * @return Return False if a link does not exists, otherwise true
- */
-static inline bool setRoot(const std::string& name, Graph& graph)
-{
-  auto& map = boost::get_property(graph, boost::graph_link_map);
-  auto found = map.find(name);
+  os << "Joints:" << std::endl;
+  for (auto j : path.second)
+    os << "  " << j << std::endl;
 
-  if (found != map.end())
-    return false;
-
-  boost::get_property(graph, boost::graph_root) = name;
-}
-
-/**
- * @brief Gets the root link name (aka. World Coordinate Frame)
- * @param graph The graph
- * @return The root link name
- */
-static inline const std::string& getRoot(const Graph& graph)
-{
-  return boost::get_property(graph, boost::graph_root);
-}
-
-/**
- * @brief Adds a link to the graph
- * @param link The link to be added to the graph
- * @param graph The graph
- * @return Return False if a link with the same name allready exists, otherwise true
- */
-static inline bool addLink(LinkPtr link, Graph& graph)
-{
-  auto& map = boost::get_property(graph, boost::graph_link_map);
-  auto found = map.find(link->getName());
-
-  if (found != map.end())
-    return false;
-
-  VertexProperty info(link);
-  Vertex v = boost::add_vertex(info, graph);
-  map[link->getName()] = std::make_pair(link, v);
-  return true;
-}
-
-/**
- * @brief Get a link in the graph
- * @param name The name of the link
- * @param graph The graph
- * @return Return nullptr if link name does not exists, otherwise a pointer to the link
- */
-static inline LinkConstPtr getLink(const std::string& name, Graph& graph)
-{
-  auto& map = boost::get_property(graph, boost::graph_link_map);
-  auto found = map.find(name);
-
-  if (found == map.end())
-    return nullptr;
-
-  return found->second.first;
-}
-
-/**
- * @brief Removes a link from the graph
- * @param name Name of the link to be removed
- * @param graph The graph
- * @return Return False if a link does not exists, otherwise true
- */
-static inline bool removeLink(const std::string& name, Graph& graph)
-{
-  auto& map = boost::get_property(graph, boost::graph_link_map);
-  auto found = map.find(name);
-
-  if (found == map.end())
-    return false;
-
-  boost::remove_vertex(found->second.second, graph);
-  map.erase(name);
-
-  return true;
-}
-
-/**
- * @brief Adds joint to the graph
- * @param joint The joint to be added
- * @param graph The graph
- * @return Return False if parent or child link does not exists and if joint name already exists in the graph, otherwise true
- */
-static inline bool addJoint(JointPtr joint, Graph& graph)
-{
-  auto& link_map = boost::get_property(graph, boost::graph_link_map);
-  auto& joint_map = boost::get_property(graph, boost::graph_joint_map);
-  auto parent = link_map.find(joint->parent_link_name);
-  auto child = link_map.find(joint->child_link_name);
-  auto found = joint_map.find(joint->getName());
-
-  if ( (parent == link_map.end()) || (child == link_map.end()) || (found != joint_map.end()) )
-    return false;
-
-  double d = joint->parent_to_joint_origin_transform.translation().norm();
-  EdgeProperty info(joint, d);
-  std::pair<Edge, bool> e = boost::add_edge(parent->second.second, child->second.second, info, graph);
-  assert(e.second == true);
-  joint_map[joint->getName()] = std::make_pair(joint, e.first);
-
-  return true;
-}
-
-/**
- * @brief Get a joint in the graph
- * @param name The name of the joint
- * @param graph The graph
- * @return Return nullptr if joint name does not exists, otherwise a pointer to the joint
- */
-static inline JointConstPtr getJoint(const std::string& name, Graph& graph)
-{
-  auto& map = boost::get_property(graph, boost::graph_joint_map);
-  auto found = map.find(name);
-
-  if (found == map.end())
-    return nullptr;
-
-  return found->second.first;
-}
-
-/**
- * @brief Removes a joint from the graph
- * @param name Name of the joint to be removed
- * @param graph The graph
- * @return Return False if a joint does not exists, otherwise true
- */
-static inline bool removeJoint(const std::string& name, Graph& graph)
-{
-  auto& map = boost::get_property(graph, boost::graph_joint_map);
-  auto found = map.find(name);
-
-  if (found == map.end())
-    return false;
-
-  boost::remove_edge(found->second.second, graph);
-  map.erase(name);
-
-  return true;
-}
-
-static inline Vertex getVertex(const std::string& name, const Graph& graph)
-{
-  auto& map = boost::get_property(graph, boost::graph_link_map);
-  auto found = map.find(name);
-
-  if (found == map.end())
-    return Vertex();
-
-  return found->second.second;
-}
-
-static inline Edge getEdge(const std::string& name, const Graph& graph)
-{
-  auto& map = boost::get_property(graph, boost::graph_joint_map);
-  auto found = map.find(name);
-
-  if (found == map.end())
-    return Edge();
-
-  return found->second.second;
+  return os;
 }
 
 } // namespace graph
